requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
google-genai
python-dotenv>=1.0.1
pypdf>=4.1.0
chromadb>=0.4.24
sentence-transformers>=2.5.1
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
typing-extensions>=4.10.0
python-multipart>=0.0.9